requirements.txt
# used
stem
requests
PySocks

# not needed anymore
# selenium
# tbselenium
